+++ migrations/20250519032653_create_withdraws_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE IF NOT EXISTS withdraws (
     id SERIAL PRIMARY KEY NOT NULL,
     user_id bigint NOT NULL,
     order_id bigint NOT NULL,
-    points bigint DEFAULT 0,
+    points NUMERIC(10, 2) DEFAULT 0,
     created_at timestamp DEFAULT now(),
     updated_at timestamp DEFAULT now()
 );
